# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API бэкенда видеоконференций.",
        "title": "Video Conference API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/meetings": {
            "post": {
                "description": "Дата ДД.ММ.ГГГГ, время ЧЧ:ММ. При пересечении слотов — 409.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meetings"
                ],
                "summary": "Создать встречу",
                "parameters": [
                    {
                        "description": "Данные встречи",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.CreateMeetingBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.CreateMeetingSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorJSON"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorJSON"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorJSON"
                        }
                    }
                }
            }
        },
        "/meetings/current": {
            "get": {
                "description": "Встреча на сегодня в интервале [start, start+duration).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meetings"
                ],
                "summary": "Текущая встреча",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID создателя",
                        "name": "creator_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.CurrentMeetingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorJSON"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorJSON"
                        }
                    }
                }
            }
        },
...
        "/meetings/upcoming": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meetings"
                ],
                "summary": "Предстоящие встречи",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID создателя (UUID)",
                        "name": "creator_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.MeetingsUpcomingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorJSON"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorJSON"
                        }
                    }
                }
            }
        },
        "/meetings/{id}": {
            "get": {
                "description": "Параметр id — суффикс после /meeting/ или /room/ в link.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meetings"
                ],
                "summary": "Встреча по коду ссылки",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Код из ссылки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.GetMeetingByCodeSuccess"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorJSON"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorJSON"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meetings"
                ],
                "summary": "Обновить встречу",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID встречи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Поля для обновления",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.UpdateMeetingBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.UpdateMeetingSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorJSON"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorJSON"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorJSON"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorJSON"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "meetings"
                ],
                "summary": "Удалить встречу",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID встречи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "creator_id",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.DeleteMeetingBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.DeleteMeetingSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorJSON"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorJSON"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorJSON"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.CreateMeetingBody": {
            "type": "object",
            "properties": {
                "creator_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "date": {
                    "type": "string",
                    "example": "15.05.2006"
                },
                "description": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer",
                    "example": 60
                },
                "invite_emails": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "link": {
                    "type": "string"
                },
                "room_id": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string",
                    "example": "14:30"
                }
            }
        },
        "controllers.CreateMeetingSuccess": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Meeting"
                },
                "invite_emails_failed": {
                    "type": "integer"
                },
                "invite_emails_sent": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "controllers.CurrentMeetingResponse": {
            "type": "object",
            "properties": {
                "meeting": {
                    "$ref": "#/definitions/models.Meeting"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "controllers.DeleteMeetingBody": {
            "type": "object",
            "properties": {
                "creator_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "controllers.DeleteMeetingSuccess": {
            "type": "object",
            "properties": {
                "cancellation_emails_failed": {
                    "type": "integer"
                },
                "cancellation_emails_sent": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "controllers.ErrorJSON": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Неверные данные запроса"
                }
            }
        },
        "controllers.GetMeetingByCodeSuccess": {
            "type": "object",
            "properties": {
                "meeting": {
                    "$ref": "#/definitions/models.Meeting"
                }
            }
        },
        "controllers.MeetingsUpcomingResponse": {
            "type": "object",
            "properties": {
                "meetings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Meeting"
                    }
                }
            }
        },
...
        "controllers.UpdateMeetingBody": {
            "type": "object",
            "properties": {
                "creator_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "link": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
        "controllers.UpdateMeetingSuccess": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Meeting"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.Meeting": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "creator_id": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "link": {
                    "type": "string"
                },
                "room_id": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}